Kernels - creation, connection
