+++ chainerx_cc/chainerx/kernels/CMakeLists.txt
@@ -1,0 +1,15 @@
+install(FILES
+    connection.h
+    creation.h
+    indexing.h
+    linalg.h
+    logic.h
+    manipulation.h
+    math.h
+    misc.h
+    normalization.h
+    pooling.h
+    sorting.h
+    statistics.h
+    DESTINATION include/chainerx/kernels
+    )
