+++ chainerx_cc/chainerx/native/native_device/conv.cc
@@ -16,13 +16,13 @@
 #include "chainerx/error.h"
 #include "chainerx/indexable_array.h"
 #include "chainerx/indexer.h"
+#include "chainerx/kernels/connection.h"
+#include "chainerx/kernels/creation.h"
 #include "chainerx/macro.h"
 #include "chainerx/native/col2im.h"
 #include "chainerx/native/im2col.h"
 #include "chainerx/native/op_regist.h"
 #include "chainerx/native/tensor_dot.h"
-#include "chainerx/routines/connection.h"
-#include "chainerx/routines/creation.h"
 #include "chainerx/routines/manipulation.h"
 #include "chainerx/shape.h"
 #include "chainerx/stack_vector.h"
