+++ chainerx_cc/chainerx/native/native_device/fill.cc
@@ -7,11 +7,11 @@
 #include "chainerx/dtype.h"
 #include "chainerx/indexable_array.h"
 #include "chainerx/indexer.h"
+#include "chainerx/kernels/creation.h"
 #include "chainerx/macro.h"
 #include "chainerx/native/data_type.h"
 #include "chainerx/native/elementwise.h"
 #include "chainerx/native/op_regist.h"
-#include "chainerx/routines/creation.h"
 #include "chainerx/routines/misc.h"
 #include "chainerx/scalar.h"
 #include "chainerx/shape.h"
