+++ chainerx_cc/chainerx/routines/connection.cc
@@ -17,6 +17,7 @@
 #include "chainerx/dims.h"
 #include "chainerx/error.h"
 #include "chainerx/graph.h"
+#include "chainerx/kernels/connection.h"
 #include "chainerx/macro.h"
 #include "chainerx/op_registry.h"
 #include "chainerx/routines/creation.h"
